api 통신에서 client와 server의 응답 logic 변경 등
- api 통신에서 client와 server의 응답 logic 변경
  - server에서 응답 객체가 불필요한 경우 비어있는 상태로 반환
  - client에서 status code에 따라 다르게 동작하는 logic을 제거
- client에서 접근(사용) 권한이 없는 사용자에 대한 1차 접근 방지를 구현
  - ex. 비로그인 사용자에게는 좋아요 버튼과 댓글 작성 버튼을 렌더링하지 않음
- 서버에서 접근 권한이 부여되지 않은 사용자에 대한 2차 접근 방지를 구현
  - ex. 비로그인 사용자가 게시물 수정 요청을 보내면 서버에서 요청 거부
- HeartBtn 컴포넌트화
- 일부 코드 주석 작성
- 서버에서 행하는 모든 코드를 try-catch문 안으로 이동
- 빈 문자열 입력 감지 코드 리팩토링
- 회원가입시 실패 또는 성공 결과 서버 요청 주소로 이동 ➔ 알람 후 홈 화면으로 이동

[#1] [#13] [#7] [#5]

+++ pages/api/heart/delete.js
@@ -4,9 +4,11 @@ import { getServerSession } from "next-auth";
 import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(request, response) {
-  const session = await getServerSession(request, response, authOptions);
   if (request.method == "DELETE") {
     try {
+      const session = await getServerSession(request, response, authOptions);
+      if (!session) return response.status(401).json();
+
       const client = await connectDB;
       const db = client.db("forum");
       const result = await db.collection("heart").deleteOne({
@@ -14,9 +16,10 @@ export default async function handler(request, response) {
         postId: new ObjectId(request.body),
       });
 
-      return response.status(200).json(result);
+      return response.status(200).json();
     } catch (error) {
-      return response.status(500).json(error);
+      console.log(error);
+      return response.status(500).json();
     }
   }
 }
